move transformer to camel Processor, tests WIP
